finally got two tables to work!
