+++ updateFciIndex.py
@@ -10,7 +10,7 @@
 
 import fciUtils
 import operator
-from database import *
+from db_models import *
 import pdb
